+++ playground/test_models.py
@@ -1,31 +1,39 @@
 import unittest
-from unittest import MagicMock
-from models import Table, Result, Links, Resolver
+from unittest.mock import MagicMock
+from models import Table, Result, Link, Resolver
 
 class TestTable(unittest.TestCase):
+    @unittest.skip("TODO")
     def test_getRawResult(self):
         pass
 
+    @unittest.skip("TODO")
     def test_getResult(self):
         pass
 
+    @unittest.skip("TODO")
     def test_resultExists(self):
         pass
 
+    @unittest.skip("TODO")
     def test_validateTable(self):
         pass
 
 class TestResult(unittest.TestCase):
+    @unittest.skip("TODO")
     def test_get(self):
         pass
 
 class TestLink(unittest.TestCase):
+    @unittest.skip("TODO")
     def test_valid(self):
         pass
 
 class TestResolver(unittest.TestCase):
-    def test_get(self)
+    @unittest.skip("TODO")
+    def test_get(self):
         pass
 
+    @unittest.skip("TODO")
     def test_update(self):
         pass